+++ HISTORY.md
@@ -43,6 +43,7 @@
 * `std::array<T, N>` is backed by a `char` tiledb attribute since the size is not guaranteed.
 * `Attribute::set_cell_val_num()` is deprecated. This is now deduced from the Attribute type.
 * Headers have the `tiledb_cpp_api_` prefix removed. For example, the include is now `#include <tiledb/attribute.h>`
+* Modified `Array::max_buffer_sizes` to always return a pair of sizes in bytes.
 
 # TileDB v1.2.1 Release Notes